+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The file format of it is based on [Keep a Changelog](http://keepachangelog.com/e
 For public Changelog covering all changes done to Pipedrive’s API, webhooks and app extensions platforms, see [public Changelog](https://pipedrive.readme.io/docs/changelog) with discussion area in [Developers Community](https://devcommunity.pipedrive.com/c/documentation/changelog/19).
 
 ## [Unreleased]
+### Added
+- Added `GET /v2/deals/installments` endpoint to fetch all installments added to a list of deals with cursor pagination
+- Added `POST /v2/deals/{id}/installments` endpoint to add an installment to a deal
+- Added `PATCH /v2/deals/{id}/installments/{installment_id}` endpoint to edit an installment
+- Added `DELETE /v2/deals/{id}/installments/{installment_id}` endpoint to delete an installment
 
 ## [24.1.1] - 2025-02-11
 ### Fixed